Optimized terrain rendering, fixed some shader rendering
Terrain now draws way less quads, fixed vertex attrib binding, added
perlin noise terrain gen

+++ source/_include/world/chunkmanager.h
@@ -30,7 +30,7 @@ public:
 
 	bool populateChunks();
 
-	void update( glm::ivec2 bottomLeft );
+	void update( glm::vec2 bottomLeft );
 
 	int getChunksRenderedX() const;
 	void setChunksRenderedX( int renderx );
